Debugging
Determined that the order of parameters in the Appendix A example code
was wrong, and fixed it. This matters because `pyproject_hooks` passes
them positionally, despite that both default to `None` and it's not at
all obvious what order they "should" be in.

I commented on https://github.com/python/peps/issues/2536, and cited
https://github.com/pypa/packaging.python.org/pull/1111, to draw
attention to the issue in the PEP. The corresponding documentation
issue, https://github.com/pypa/packaging.python.org/issues/955, still
appears not to be resolved.