Add basic metadata
A minimal PKG-INFO, compliant with version 2.2 of the core metadata
specification (as in turn required for a PEP 517-compliant,
pyproject.toml-based sdist) is automatically generated in memory and
added directly to the output sdist. It specifies the minimum information
required by the specification - name, version and metadata version.

It may become necessary to expand this information later.

The wheel building process will assume it may blindly copy PKG-INFO from
the sdist to METADATA in the .dist-info directory of the wheel.