Add functionality to read pyproject.toml
Rather than vendoring `tomli` for Python before 3.11, `bbbb` simply
specifies it as a build requirement using the appropriate hooks. `tomli`
itself has no dependencies, so there is no risk of a dependency cycle.

No attempt is made, nor shall be made, to validate pyproject.toml
contents (nor existence; though existence should be assured by the fact
that `bbbb` was invoked). Exceptions simply propagate. Such validation
is beyond the scope of a minimal build backend that is trying to be a
*pure* backend that is as small as possible (modulo reasonably good
coding style and basic quality of life features).