Reorganization
Code specific to building wheels is moved to `wheel.py`, which also has
some stubs for the sdist hooks. Imports are also trimmed appropriately.

The idea is that the project will eventually build wheels and sdists for
a full "development" version, but also wheels that only contain the
wheel-building part of the backend - for use in already-prepared sdists
that have been downloaded by an installer. This further minimizes the
size requirements for a build system when someone else publishes an
sdist that expects to use `bbbb`.

(Since `bbbb` intends to support projects with compiled extensions, it's
worthwhile to think about this situation. Other projects using `bbbb`
may not be able to provide wheels for all platforms.)

No sdist should be shipped for the wheel-building wheel, because it's
only ever intended to be used as an automatically-downloaded, miniature
build system. Needing to build it would defeat the purpose. However,
wheels might eventually be created that pre-compile to .pyc.

It's also intended that the wheel-building wheel will provision a
single module, rather than a package containing that module. Similarly,
the metadata in said wheel should be kept to the minimum required by the
relevant specifications (making a friendly description for PyPI is not
a priority), while the "main" project will have full metadata.

This setup was also tested by temporarily changing the `[build-system]`
in `pyproject.toml` to `bbbb.wheel`. This raises the exception demanded
by the PEP 517 specification when attempting to build an sdist, and
works properly when attempting to build a wheel - module the fact that
the source tree hasn't been filtered by the sdist-preparation code.