Use example sdist filter hook
A hook is implemented which excludes the `dist/` folder created by
`build`, as well as any `__pycache__` folders, dotfiles, and the hook
file itself.

Note that the PYTHONDONTWRITEBYTECODE environment variable is still
needed when using `build` to build the self-hosting wheel for this
project, to exclude `.pyc` files (and `__pycache__` folders) from the
wheel. This is because, after `build` unpacks the sdist to build a wheel
from it, it will use the code *from the unpacked sdist* (not the
original source tree) to run `bbbb` and build the wheel. (This is build
isolation working as intended.)