Initial tests
The code itself is imported using a `sys.path` hack so that it doesn't
have to be installed first. Currently editable installs aren't
supported, so changes in the codebase wouldn't get tested. But even once
supported, the editable install functionality should be tested
separately and explicitly.

This initial test simply copies the working copy into the test
directory, and then verifies that building an sdist works without error
and creates the destination folder if not present.