Improved filter for self-build
There is now a test directory which should be excluded from the build
process, and testing the self-build process creates a test_sdist
directory which should be excluded from the test-self-built sdist.
Updated the sdist filter (which is both a demonstration for testing, and
part of the project's build process) accordingly.