Use `build` frontend in testing
This way we don't need to work through the PEP 517 API directly and get
the subprocess hooks etc. for free. Notably, this means we don't need to
use a `sys.path` hack to get at the current `bbbb` source code - in
fact, we use the code from the copied pytest environment (thanks to the
`backend-path` specification in `pyproject.toml`).

Also, notably, this emulates the *recommended* workflow for actual use.

The programmatic API for `build` doesn't directly expose the "build
wheel from sdist" logic, so we'll still need a test flow that unpacks
the sdist within the test environment and then builds a wheel from it.
But since this is happening within a temporary folder already created
by Pytest, we don't need to do any cleanup.