Add initial tests for wheel building
These tests are broken because the manifest file for the sdist is
reused, but the wheel is supposed to contain different files. In fact,
the current setup won't work, because the manifest file parsing
currently assumes a single top-level folder with a hard-coded name.
Unfortunately it will be necessary to undo the simplification there and
force the manifest files to list the complete paths, at least for now.