Fix wheel tests by including a wheel manifest
For now, we assume that the resulting wheels will *not* include .pyc
files or __pycache__ directories, which we can ensure by clearing them
from the working tree and then setting PYTHONDONTWRITEBYTECODE=1 before
running the tests. (This matches how the published wheels for the
project have been created so far.)

In the future we'll want to allow more flexibility here.