New test configuration
A single TOML file is used to store information for testing: the
expected manifest for both sdist and wheel results, and hashes and file
sizes from RECORD. Placeholders for the project (not package!) name and
version can be substituted in, using vanilla `str.format`.

The file format also specially indicates the path to the RECORD file in
the wheel - this may not be necessary; TBD.

The expected name and version should be read from the original
`pyproject.toml` from the test data. This is less maintenance overall;
in particular, it will mean the self-test data won't need to be updated
simply because the project version number was changed.

The tests have been updated to parse and use this file and thus verify
the sdist and wheel file manifests.

Future tests will also verify:

* The sdist's `pyproject.toml` specifies a matching name and version
* The wheel's `RECORD` specifies the correct hashes and file sizes (this
  also implicitly verifies file contents) and has the expected
  self-referential line

(Eventually the hash/file size data will be removed from the self-test
so that it doesn't have to be updated every time the project code
changes)