More reorganization
This structure makes it feasible to include tests in both sdists and
wheels for `bbbb-dev`. Note that since this isn't yet optional, all
tests currently fail (the expected output for the self-test won't
include tests).

* The "core" package is now just a top-level `bbbb` module. Instead of
  raising an exception when asked to build an sdist, it retrieves the
  dev package (using a PEP 517 hook) and delegates to it.

* The "dev" package is `bbbb_dev`, which has an `__init__.py` (so the
  import name is `bbbb_dev`; convenient but not very important) and a
  `tests` subfolder. It can build an sdist directly and a wheel by using
  its static import of `bbbb`.

This way, users can have the best of both worlds by specifying `bbbb` as
the backend in `pyproject.toml`: when they request a wheel, `bbbb` does
its thing normally, and when they request an sdist, `bbbb-dev` is set up
in the build environment, but the resulting sdist's `pyproject.toml`
asks for `bbbb` - so *their* users don't need the whole package.

On the other hand, it's still valid to use `[bbbb-dev]` for
`build-system.requires` and `"bbbb_dev"` for
`build-system.build-backend`, since a `bbbb-dev` installation will still
include `bbbb`.