Redesign "config" interface
* The config data produced by `_get_config` is now specific to the kind of
  build; a single `'toml'` key is used for config from the corresponding
  section of `pyproject.toml` - that is, wheel builds don't see sdist
  config and vice-versa.

* Any custom options provided on the command line are included under the
  `'commandline'` key, so user build hooks can parse and use them.
  Instead of parsing out specific user keys for wheel tags, any key may
  be used and they have no special meaning to `bbbb`.

* The wheel tags are now stored within wheel config: therefore they're
  only present when necessary, and can be overridden with the
  `[tool.bbbb.wheel.tags]` section of `pyproject.toml`.

It was unrealistic to expect wheel tags to be provided on the command
line in normal cases; the user wheel build hook (not yet integrated)
needs to do the computation, and it may take the command line into
consideration in doing so. It's also useful to document in
`pyproject.toml` if the project can only ever create one kind of wheel
(and that happens to differ from `py3-none-any`).