Add option to exclude tests
Multiple fixes for the sdist filter:

* It looks for tests in the right place.

* It checks command-line config and uses a separate function to do the
  main work after verifying that config. For sanity, any unrecognized
  or ill-formed command-line options are rejected.

* When the `build` command line specifies `-C exclude-tests` (which is
  equivalent to passing `{'exclude-tests': ''}` in the API), the filter
  will exclude tests; otherwise the tests are included.

TODO: make it so that the filter is only invoked once with a generator
of file paths or something.

Currently, wheel tests fail because the direct-wheel path can't actually
exclude tests, and the wheel-via-sdist path is set up to include them
(because it's supposed to verify that the wheel can filter things that
the sdist included).