Reorganize test environments
When an environment is set up to test wheel building, the actual project
is now put into a `project` subdirectory, while the `test_dist` will be
parallel to that. This allows for test projects that don't need to
filter out the `test_dist` directory.

While using `pyproject-build` from the command line, the `--outdir`
switch can be used to control where wheels and sdists are placed (and
`bbbb` correctly respects this designation). Therefore this testing
setup doesn't hide any bugs, only a small limitation. (We aren't testing
`build` itself, only `bbbb`, so tests will not play around with this.)

This is necessary to make other tests pass without additional effort
that might not reflect real use cases.